// ClienteApp/src/app/logica-apis/model/compromissosModel.ts
/**
 * API Legislação
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface CompromissosModel {

    _id?: string;
    
    titulo?: string;

    dataInicio?: Date;

    dataFim?: Date;

    dataCriacao?: Date;

    dataAtualizacao?: Date;

    status?: string;

    descricao?: string;
}